fix(kclipboardprovider): refactor to script setups
