+++ src/components/KClipboardProvider/KClipboardProvider.vue
@@ -1,0 +1,14 @@
+<template>
+  <slot :copy-to-clipboard="copyTextToClipboard" />
+</template>
+
+<script setup lang="ts">
+import { copyTextToClipboard } from '@/utilities/copyTextToClipboard'
+
+import type { ClipboardProviderSlots } from '@/types/clipboard-provider'
+
+defineOptions({
+  name: 'KClipboardProvider',
+})
+defineSlots<ClipboardProviderSlots>()
+</script>
